refactor(practice-template): move service to services/ and adjust imports
